backend/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
clip==0.2.0
cloudinary==1.41.0
colorama==0.4.6
diffusers==0.31.0
Django==5.1.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
Flask==3.0.3
fsspec==2024.10.0
ftfy==6.3.1
gradio==5.7.1
gradio_client==1.5.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.26.3
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
mysqlclient==2.2.4
networkx==3.4.2
numpy==2.1.3
openai-clip==1.0.1
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==10.4.0
pydantic==2.10.2
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
replicate==1.0.4
requests==2.32.3
respond==1.0
rich==13.9.4
ruff==0.8.1
safehttpx==0.1.1
safetensors==0.4.5
scipy==1.14.1
semantic-version==2.10.0
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.41.3
sympy==1.13.1
tokenizers==0.20.3
tomlkit==0.12.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.46.3
typer==0.14.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
waitress==3.0.2
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.4
zipp==3.21.0
